Move businesslogic from API to service
