+++ src/main/kotlin/no/nav/syfo/melding/api/MeldingApi.kt
@@ -9,7 +9,6 @@ import no.nav.syfo.application.api.checkVeilederTilgang
 import no.nav.syfo.melding.MeldingService
 import no.nav.syfo.client.veiledertilgang.VeilederTilgangskontrollClient
 import no.nav.syfo.domain.PersonIdent
-import no.nav.syfo.melding.domain.MeldingType
 import no.nav.syfo.util.*
 import java.util.UUID
 
@@ -117,24 +116,13 @@ fun Route.registerMeldingApi(
             )
 
             val meldingUuid = call.meldingUuid()
-            val innkommendeLegeerklaring = meldingService.getMeldingFraBehandler(meldingUuid)
-                ?.takeIf { it.type == MeldingType.FORESPORSEL_PASIENT_LEGEERKLARING }
-                ?: throw IllegalArgumentException("Failed to create retur av legeerklæring: Melding with uuid $meldingUuid does not exist")
-            val opprinneligForesporselLegeerklaring = meldingService.getUtgaendeMeldingerInConversation(
-                conversationRef = innkommendeLegeerklaring.conversationRef,
-                personIdent = innkommendeLegeerklaring.arbeidstakerPersonIdent,
-            ).first { it.type == MeldingType.FORESPORSEL_PASIENT_LEGEERKLARING }
-
             val requestDTO = call.receive<ReturAvLegeerklaringRequestDTO>()
-            val returAvLegeerklaring = requestDTO.toMeldingTilBehandler(
-                opprinneligForesporselLegeerklaring = opprinneligForesporselLegeerklaring,
-                innkommendeLegeerklaring = innkommendeLegeerklaring,
-                veilederIdent = veilederIdent,
-            )
 
             meldingService.createReturAvLegeerklaring(
                 callId = getCallId(),
-                returAvLegeerklaring = returAvLegeerklaring,
+                meldingUuid = meldingUuid,
+                veilederIdent = veilederIdent,
+                document = requestDTO.document,
             )
 
             call.respond(HttpStatusCode.OK)
