+++ src/test/kotlin/no/nav/syfo/melding/api/MeldingApiPostSpek.kt
@@ -399,6 +399,8 @@ class MeldingApiPostSpek : Spek({
                                     NAV_PERSONIDENT_HEADER,
                                     personIdent.value
                                 )
+                                addHeader(HttpHeaders.ContentType, ContentType.Application.Json.toString())
+                                setBody(objectMapper.writeValueAsString(returDTO))
                             }
                         ) {
                             response.status() shouldBeEqualTo HttpStatusCode.BadRequest
@@ -428,6 +430,8 @@ class MeldingApiPostSpek : Spek({
                                     NAV_PERSONIDENT_HEADER,
                                     personIdent.value
                                 )
+                                addHeader(HttpHeaders.ContentType, ContentType.Application.Json.toString())
+                                setBody(objectMapper.writeValueAsString(returDTO))
                             }
                         ) {
                             response.status() shouldBeEqualTo HttpStatusCode.BadRequest
